test: optimizer bug
