+++ test/compile/data/optimizers.test.ts
@@ -1,12 +1,11 @@
 import {DataFlowNode} from '../../../src/compile/data/dataflow';
+import {ParseNode} from '../../../src/compile/data/formatparse';
 import {ImputeNode} from '../../../src/compile/data/impute';
-import {optimizeDataflow} from '../../../src/compile/data/optimize';
-import {MergeIdenticalNodes, MergeTimeUnits} from '../../../src/compile/data/optimizers';
+import {MergeIdenticalNodes, MergeParse, MergeTimeUnits} from '../../../src/compile/data/optimizers';
+import {TimeUnitComponent, TimeUnitNode} from '../../../src/compile/data/timeunit';
 import {Transform} from '../../../src/transform';
-import {parseLayerModel} from '../../util';
-import {FilterNode} from './../../../src/compile/data/filter';
-import {TimeUnitNode, TimeUnitComponent} from '../../../src/compile/data/timeunit';
 import {hash} from '../../../src/util';
+import {FilterNode} from './../../../src/compile/data/filter';
 
 describe('compile/data/optimizer', () => {
   describe('mergeIdenticalNodes', () => {
@@ -20,8 +19,10 @@ describe('compile/data/optimizer', () => {
       const root = new DataFlowNode(null, 'root');
       const transform1 = new ImputeNode(root, transform);
       new ImputeNode(root, transform);
+
       const optimizer = new MergeIdenticalNodes();
       optimizer.run(root);
+
       expect(root.children).toHaveLength(1);
       expect(root.children[0]).toEqual(transform1);
       expect(optimizer.mutatedFlag).toEqual(true);
@@ -41,8 +42,10 @@ describe('compile/data/optimizer', () => {
       const transform3 = new FilterNode(root, null, 'datum.x > 2');
 
       new FilterNode(root, null, 'datum.x > 2');
+
       const optimizer = new MergeIdenticalNodes();
       optimizer.run(root);
+
       expect(root.children).toHaveLength(2);
       expect(root.children).toEqual([transform1, transform3]);
     });
@@ -64,6 +67,7 @@ describe('compile/data/optimizer', () => {
       expect(parent.children).toHaveLength(2);
       expect(a.children).toHaveLength(2);
       expect(b.children).toHaveLength(2);
+
       const optimizer = new MergeIdenticalNodes();
       optimizer.mergeNodes(parent, [a, b]);
       optimizer.setMutated();
@@ -75,35 +79,6 @@ describe('compile/data/optimizer', () => {
       expect(a2.parent).toBe(a);
       expect(b1.parent).toBe(a);
       expect(b2.parent).toBe(a);
-    });
-  });
-
-  describe('MergeBins', () => {
-    it('should rename signals when merging BinNodes', () => {
-      const transform = {
-        bin: {extent: [0, 100], anchor: 6},
-        field: 'Acceleration',
-        as: ['binned_acceleration_start', 'binned_acceleration_stop']
-      };
-      const model = parseLayerModel({
-        layer: [
-          {
-            transform: [transform],
-            mark: 'rect',
-            encoding: {}
-          },
-          {
-            transform: [transform],
-            mark: 'rect',
-            encoding: {}
-          }
-        ]
-      });
-      model.parse();
-      optimizeDataflow(model.component.data, model);
-      expect(model.getSignalName('layer_0_bin_extent_0_100_anchor_6_maxbins_10_Acceleration_bins')).toEqual(
-        'layer_1_bin_extent_0_100_anchor_6_maxbins_10_Acceleration_bins'
-      );
     });
   });
 
@@ -144,6 +119,56 @@ describe('compile/data/optimizer', () => {
         {as: 'c_yr', expr: 'datetime(year(datum["c"]), 0, 1, 0, 0, 0, 0)', type: 'formula'},
         {as: 'b_yr', expr: 'datetime(year(datum["b"]), 0, 1, 0, 0, 0, 0)', type: 'formula'}
       ]);
+    });
+  });
+
+  describe('MergeParse', () => {
+    it('should merge non-conflicting ParseNodes', () => {
+      const root = new DataFlowNode(null, 'root');
+      const parse1 = new ParseNode(root, {a: 'number', b: 'string'});
+      new ParseNode(root, {b: 'string', c: 'boolean'});
+
+      const optimizer = new MergeParse();
+      optimizer.run(parse1);
+
+      expect(root.children.length).toEqual(1);
+      const mergedParseNode = root.children[0] as ParseNode;
+      expect(mergedParseNode.parse).toEqual({a: 'number', b: 'string', c: 'boolean'});
+    });
+
+    it('should not merge conflicting ParseNodes', () => {
+      const root = new DataFlowNode(null, 'root');
+      const parse1 = new ParseNode(root, {a: 'number', b: 'string'});
+      new ParseNode(root, {a: 'boolean', d: 'date'});
+      new ParseNode(root, {a: 'boolean', b: 'string'});
+
+      const optimizer = new MergeParse();
+      optimizer.run(parse1);
+
+      expect(root.children.length).toEqual(1);
+      const mergedParseNode = root.children[0] as ParseNode;
+      expect(mergedParseNode.parse).toEqual({b: 'string', d: 'date'});
+      const children = mergedParseNode.children as [ParseNode, ParseNode];
+      expect(children[0].parse).toEqual({a: 'number'});
+      expect(children[1].parse).toEqual({a: 'boolean'});
+    });
+
+    it('should merge when there is no parse node', () => {
+      const root = new DataFlowNode(null, 'root');
+      const parse = new ParseNode(root, {a: 'number', b: 'string'});
+      const parseChild = new DataFlowNode(parse);
+      const otherChild = new DataFlowNode(root);
+
+      const optimizer = new MergeParse();
+      optimizer.run(parse);
+
+      expect(root.children.length).toEqual(1);
+      const mergedParseNode = root.children[0] as ParseNode;
+      expect(mergedParseNode.parse).toEqual({a: 'number', b: 'string'});
+      const children = mergedParseNode.children;
+      expect(children).toHaveLength(2);
+      expect(children[0]).toEqual(otherChild);
+      expect(children[1]).toEqual(parseChild);
     });
   });
 });
